+++ internal/provider/data_source_registry_module.go
@@ -96,6 +96,8 @@ type modelTFEVCSRepo struct {
 	GHAInstallationID types.String `tfsdk:"github_app_installation_id"`
 	RepositoryHTTPURL types.String `tfsdk:"repository_http_url"`
 	ServiceProvider   types.String `tfsdk:"service_provider"`
+	SourceDirectory   types.String `tfsdk:"source_directory"`
+	TagPrefix         types.String `tfsdk:"tag_prefix"`
 	Tags              types.Bool   `tfsdk:"tags"`
 	TagsRegex         types.String `tfsdk:"tags_regex"`
 	WebhookURL        types.String `tfsdk:"webhook_url"`
@@ -127,6 +129,8 @@ func modelFromTFEVCSRepo(v *tfe.VCSRepo) modelTFEVCSRepo {
 		GHAInstallationID: types.StringValue(v.GHAInstallationID),
 		RepositoryHTTPURL: types.StringValue(v.RepositoryHTTPURL),
 		ServiceProvider:   types.StringValue(v.ServiceProvider),
+		SourceDirectory:   types.StringValue(v.SourceDirectory),
+		TagPrefix:         types.StringValue(v.TagPrefix),
 		Tags:              types.BoolValue(v.Tags),
 		TagsRegex:         types.StringValue(v.TagsRegex),
 		WebhookURL:        types.StringValue(v.WebhookURL),